addressing oxy_fitting
