+++ ctdcal/tests/test_oxy_fitting.py
@@ -4,7 +4,8 @@ from pathlib import Path
 import numpy as np
 import scipy
 
-from ctdcal import oxy_fitting
+from ctdcal.fitting.fit_oxy import calculate_weights
+from ctdcal.processors.proc_oxy_odf import gather_oxy_params
 
 
 def test_gather_oxy_params(caplog, tmp_path):
@@ -13,7 +14,7 @@ def test_gather_oxy_params(caplog, tmp_path):
     # breakpoint()
     assert not Path(tmp_path / "90909").exists()
     with caplog.at_level(logging.INFO):
-        oxy_params = oxy_fitting.gather_oxy_params(f_path)
+        oxy_params = gather_oxy_params(f_path)
         assert "Failed to load" in caplog.messages[0]
         assert oxy_params.isnull().values.all()
 
@@ -37,6 +38,6 @@ def test_calculate_weights():
     w_bins = [20, 20, 25, 25, 50, 50, 100, 100, 200, 200, 500, 500]
     wgt_manual = scipy.interpolate.interp1d(p_bins, w_bins)(pressure)   #   Pull weights out
 
-    wgt = oxy_fitting.calculate_weights(pressure)
+    wgt = calculate_weights(pressure)
 
     assert np.array_equal(wgt, wgt_manual)
